refactor: add `api/illustration` folder
