+++ components/api/illustration/useDeleteIllustration.js
